heartbeat: refactor heartbeat channel usage
